# Conflicts:
#	common/src/main/java/com/st0x0ef/stellaris/client/screens/components/Gauge.java
#	common/src/main/java/com/st0x0ef/stellaris/client/skies/record/CustomVanillaObject.java
#	common/src/main/java/com/st0x0ef/stellaris/client/skies/renderer/SkyRenderer.java
#	common/src/main/java/com/st0x0ef/stellaris/common/entities/vehicles/RocketEntity.java
#	common/src/main/java/com/st0x0ef/stellaris/common/items/CanItem.java

+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/CoalGeneratorScreen.java
@@ -57,7 +57,7 @@ public class CoalGeneratorScreen extends AbstractContainerScreen<CoalGeneratorMe
             return;
         }
 
-        energyGauge.update(blockEntity.getWrappedEnergyContainer().getStoredEnergy());
+        energyGauge.update((int)blockEntity.getWrappedEnergyContainer().getStoredEnergy());
     }
 
     @Override
