+++ common/src/main/java/com/st0x0ef/stellaris/client/screens/WaterSeparatorScreen.java
@@ -43,19 +43,19 @@ public class WaterSeparatorScreen extends AbstractContainerScreen<WaterSeparator
         }
 
         FluidTank ingredientTank = blockEntity.getIngredientTank();
-        ingredientTankGauge = new Gauge(leftPos + 50, topPos + 58, 76, 42, Component.translatable("stellaris.screen.water"), GUISprites.WATER_OVERLAY, GUISprites.WATER_SEPARATOR_OVERLAY, ingredientTank.getAmount(), ingredientTank.getMaxCapacity()-1);
+        ingredientTankGauge = new Gauge(leftPos + 50, topPos + 58, 76, 42, Component.translatable("stellaris.screen.water"), GUISprites.WATER_OVERLAY, GUISprites.WATER_SEPARATOR_OVERLAY, (int)ingredientTank.getAmount(), (int)ingredientTank.getMaxCapacity()-1);
         addRenderableWidget(ingredientTankGauge);
 
         FluidTank resultTank1 = blockEntity.getResultTanks().getFirst();
-        resultTank1Gauge = new Gauge(leftPos + 22, topPos + 54, 12, 46, Component.translatable("stellaris.screen.hydrogen"), GUISprites.HYDROGEN_OVERLAY, GUISprites.LIQUID_TANK_OVERLAY, resultTank1.getAmount(), resultTank1.getMaxCapacity()-1);
+        resultTank1Gauge = new Gauge(leftPos + 22, topPos + 54, 12, 46, Component.translatable("stellaris.screen.hydrogen"), GUISprites.HYDROGEN_OVERLAY, GUISprites.LIQUID_TANK_OVERLAY, (int)resultTank1.getAmount(), (int)resultTank1.getMaxCapacity()-1);
         addRenderableWidget(resultTank1Gauge);
 
         FluidTank resultTank2 = blockEntity.getResultTanks().get(1);
-        resultTank2Gauge = new Gauge(leftPos + 142, topPos + 54, 12, 46, Component.translatable("stellaris.screen.oxygen"), GUISprites.OXYGEN_OVERLAY, GUISprites.LIQUID_TANK_OVERLAY, resultTank2.getAmount(), resultTank2.getMaxCapacity()-1);
+        resultTank2Gauge = new Gauge(leftPos + 142, topPos + 54, 12, 46, Component.translatable("stellaris.screen.oxygen"), GUISprites.OXYGEN_OVERLAY, GUISprites.LIQUID_TANK_OVERLAY, (int)resultTank2.getAmount(), (int)resultTank2.getMaxCapacity()-1);
         addRenderableWidget(resultTank2Gauge);
 
         EnergyContainer energyContainer = blockEntity.getWrappedEnergyContainer();
-        energyGauge = new Gauge.SidewayGauge(leftPos + 64, topPos + 24, 47, 13, Component.translatable("stellaris.screen.energy"), GUISprites.SIDEWAYS_ENERGY_FULL, GUISprites.SIDEWAYS_ENERGY_OVERLAY, (int) energyContainer.getStoredEnergy(), (int) energyContainer.getMaxCapacity());
+        energyGauge = new Gauge.SidewayGauge(leftPos + 64, topPos + 24, 47, 13, Component.translatable("stellaris.screen.energy"), GUISprites.SIDEWAYS_ENERGY_FULL, GUISprites.SIDEWAYS_ENERGY_OVERLAY, (int)energyContainer.getStoredEnergy(), (int)energyContainer.getMaxCapacity());
         addRenderableWidget(energyGauge);
     }
 
@@ -69,10 +69,10 @@ public class WaterSeparatorScreen extends AbstractContainerScreen<WaterSeparator
             return;
         }
 
-        ingredientTankGauge.update(blockEntity.getIngredientTank().getAmount());
-        resultTank1Gauge.update(blockEntity.getResultTanks().getFirst().getAmount());
-        resultTank2Gauge.update(blockEntity.getResultTanks().getLast().getAmount());
-        energyGauge.update(blockEntity.getWrappedEnergyContainer().getStoredEnergy());
+        ingredientTankGauge.update((int)blockEntity.getIngredientTank().getAmount());
+        resultTank1Gauge.update((int)blockEntity.getResultTanks().getFirst().getAmount());
+        resultTank2Gauge.update((int)blockEntity.getResultTanks().getLast().getAmount());
+        energyGauge.update((int)blockEntity.getWrappedEnergyContainer().getStoredEnergy());
     }
 
     @Override
